Modify UDP probe to allow intervals shorter than timeout.
Modify UDP probe to allow intervals shorter than timeout, for example 200ms interval with 1s timeout. To make that possible, do the metrics calculation of packets sent and received in batches and skip the packets that are from the last timeout period.

Also, in the process:
= Simplify the metrics management a little bit. Don't use probeutils.StatsKeeper(). It's not providing much value in this case.
= Add more tests to cover cases where interval is shorter than timeout and fix couple of issues in test UDP server.

ORIGINAL_AUTHOR=Manu Garg <manugarg@gmail.com>
PiperOrigin-RevId: 186165323