rtcservice: Don't try to list during initialization.
List can fail for many reasons, which don't have anything to do with the configuration problems. We should either examine the List() error to find out the actual cause of failure or just not List() during init(). I am going with the latter. I have added a TODO to explore the former.

PiperOrigin-RevId: 202541985