Workaround for high loss rate during restarts in udp listener probe
When the probe binary restarts, it has to initialize a new flow state (srcHost, srcPort, seqnum) where seqnum=0. The first packet for each socket is not marked as "success" and in configs where we have hundreds of sockets, and the transmitter sends one packet/sec, we can have minute-long periods where loss rate is 100%.

Workaround is to treat the first packet of a socket (flow state seqnum==0) as a success.

This change also adds an explicit fix for the case where the sender restarts (msg.Seq() == 1). This message is also treated as a success and the flow state is reset.

PiperOrigin-RevId: 230969471