Move stats_export_interval definition to the outer layer (probe level config, rather than probe-type level config).
During the transition phase, keep the field in the probe-type config messages but log a warning.

Another notable change is that stats_export_interval doesn't have a static default value anymore. Instead, the default value is now determined based on the other fields like interval and timeout. All of that logic is in options.go.

PiperOrigin-RevId: 252467597