Provide a way to access instance labels in cloudprober config.
In cloudprober.cfg files there is an option to refer GCE instances's attributes such as zone and project by writing ".zone" and ".project" respectively. This CL adds an option to refer to instance's labels in a similar way. For example, if an instance has a shard label, its value can be accessed in the config through the variable ".labels_shard".

PiperOrigin-RevId: 259528314