Allow multiple instances of cloudprober to run sequentially.
This is most useful for tests in which a full cloudprober is
required. Each test initializes a new instance, calling
InitFromConfig and Start and when the context passed to Start is done,
the cloud prober instance will shut down fully. There is currently no
good way to ensure that the last instance has shut down before starting
a new one, so I used a time.Sleep.

- Added ClearDefaultGRPCServer() to runconfig to clear the saved gRPC
server
- On context close, Start will fully clean up the global variables found
in cloudprober.go
- Added test that starts CloudProber, stops it, and starts it again.
Before this code change, that would have always failed with the second
config ignored.