Set Content-Length header on outgoing requests
This avoids Transfer-Encoding: chunked, which the Go HTTP client
otherwise adds for us by default because it can't figure out the length
of the body.

Chunked transfer encoding is harder to handle on some servers, e.g.
wsgi+Django. It's also pretty unnecessary when we actually know the
exact length of the body in advance.