Implement context handling in serverutils
This gives us two useful behaviours: we can cancel the context and
wait for all probes to exit before cleaning up global resources, and
we can propagate the probe timeout into the probe function, so that it
can be respected by libraries that we call. The bigquery contrib probe
is updated to reflect this, so that the network calls it makes to
bigquery will be abandoned when the probe timeout is hit.