[surfacers.stackdriver] More fixes for Google Cloud Run.
See #257 for the background.

Cloud Run starts/restarts the container multiple with the same revision
id. This creates a problem for the stackdriver metrics (at least in the
short term) as for the cumulative metrics, value startTime is tied to the
Cloudprober start time and endTime cannot be before the startTime. This
condition is violated if there are multiple processes exporting data to
the same resource.