[surfacers.stackdriver] Use random task ID for Google Cloud Run. (#267)
See #257 for the background.

Cloud Run starts/restarts the container multiple with the same revision id. This creates a problem for the Stackdriver metrics (at least in the short term) as for the cumulative metrics, a point's startTime is tied to the Cloudprober start time and its endTime cannot be before the startTime. This condition is violated if there are multiple processes exporting data to the same resource.