[oauth] Fix a bug in http_token.
We were using a buffered reader for the HTTP body, which meant that once
the request body is read, it will be empty unless we reset it.

Currently using the same type of reader as the one used in the
probes/http package. I'll eventually move this to a common package.