[metrics] Improve Maps's implementation using generics. (#444)
Map implementation was complex and expensive primarily to deal with both
int64 and float64 values. At that time we used an interface "NumValue" as map
values to support both types. That's kind of unfortunate as Map is almost always
used as a counter (e.g. resp-code, validation failure map), except for external
probe output parsing where we can encounter a float map.

Now that Go supports generics, we can implement Maps using generics, with a
union of int64 and float64 (int64 | float64) as type parameter.

* New implementation improves Map's cloning performance by 65%:
-  BenchmarkMapClone    591   1890373 ns/op  2966837 B/op	 28412 allocs/op
+ BenchmarkMapClone   1506   804454 ns/op   1375970 B/op	  4153 allocs/op

* Make stackdriver surfacer implementation more streamlined and improve testing. We
   actually had a bug in the implementation earlier where we were exporting floats as int64. 

* Add more tests for prometheus surfacer, metrics.Float, and metrics.Map.