Support file glob when including config files
Allow supplying a filepath glob pattern inside the `include` so that multiple files can be included
using a single statement. Backward compatibility of including a single file is still supported as
the glob match will just return a single file path in the slice.

The glob pattern syntax is that of `filepath.Match()`.