[probes.http] Stop using the default http transport (#873)
To address #872

Currently if the disable_http2 option is set to true, http probes would fail with malformed response.

This is due to transport.ForceAttemptHTTP2 being set to false and transport.TLSNextProto being set to an empty map, BUT transport.TLSClientConfig.NextProtos being populated with "h2", "http/1.1". In other words, we wanted to disable http2, but the TLS ALPN protocols contain h2. This is not a valid configuration.

To prevent ALPN protocols from being populated, we shouldn't use the default http transport.