[probes.grpc] Provide an option to not reuse connections.
Note that in gRPC, it's not default because gRPC application usually
behave that way -- they establish a connection once and keep on reusing
it. gRPC probes try to mimic that. In a way they are more focussed on
testing the service functionality rather than connectivity.

However, if your service is behind a loadbalancer, you might want to
target as many instances as possible. To do that, its necessary that you
either establish a new connection everytime or you use a large number
of connections (num_conns) in each probe cycle.