[probes.browser] First check-in for advanced web version (#1074)
This still needs a lot of work for the full implementation, but this version is already quite useful as it lists all the timestamp directories on the top page (see screenshot in the comment below). It'll now be easy to add other features like filtering by time range.

One drawback of this current design is that it shows all timestamped directories on one page (instead of showing them by days), which may be hard to browse. I'll check this in and follow-up with filtering by date and time in the next PR.

See https://github.com/cloudprober/cloudprober/issues/958 for more context.