fix: fix sockaddr for ipv4 source addresses in ping probe
Fixes a bug in the sockaddr function where configuring a specific IPv4
source_ip resulted in the probe binding to 0.0.0.0 (all interfaces)
instead of the requested address.

The `net.IP` is a 16-byte slice, IPv4 mapped IPv6 in the case of this
bug. The previous implementation used `copy()` to transfer the slice
into the 4-byte `syscall.SockaddrInet4.Addr` array, however `copy()`
only transfers the minimum length, which in this case is the 4-byte
`SockaddrInet4.Addr`. The resulting `sa.Addr` is the first 4-bytes of
the `sourceIP (net.IP)` which is just `0.0.0.0`. So IPv4 source
addresses were being ignored, and `0.0.0.0` was being bound to.

Fix is to get respective byte ranges for IPv4 and IPv6 before the copy.